Squashed commit of the following:

commit 5bdf7e612494cbc24c0800b3d33d828b2d493507
Author: rrp <rrp@cisco.com>
Date:   Thu Jan 28 16:26:43 2016 -0500

    Remaining cosmetic changes based on code review comments for DCE checkin

commit 48770da057b8e156d183658d94d05b37b75c297a
Author: rrp <rrp@cisco.com>
Date:   Thu Jan 28 13:05:19 2016 -0500

    Get DCE-RPC SO library to build - clean with uncrustify

commit 684304c664210c1f14556e1c07100eaf166d8f58
Author: rrp <rrp@cisco.com>
Date:   Thu Jan 28 12:46:52 2016 -0500

    Get DCE-RPC SO library to build

+++ src/service_inspectors/dce_rpc/dce2_smb_module.h
@@ -62,7 +62,7 @@ struct dce2SmbProtoConf
     uint8_t smb_max_compound;
     uint16_t smb_valid_versions_mask;
     dce2SmbFileInspection smb_file_inspection;
-    int64_t smb_file_depth;
+    int16_t smb_file_depth;
     DCE2_List* smb_invalid_shares;
 };
